+++ modules/container_definition/secrets.tf
@@ -1,13 +1,17 @@
 locals {
-  ssm_prefix = "/aws/reference/secretsmanager/"
-
   secrets_keys        = keys(var.secrets)
   sorted_secrets_keys = sort(local.secrets_keys)
+
   final_secrets = [
     for key in local.sorted_secrets_keys :
     {
       name      = key
-      valueFrom = "${local.ssm_prefix}${lookup(var.secrets, key)}"
+      valueFrom = module.secret_references.valuesFrom[key]
     }
   ]
+}
+
+module "secret_references" {
+  source  = "../secret_references"
+  secrets = var.secrets
 }