+++ modules/nginx/apigw/data.tf
@@ -2,8 +2,5 @@ locals {
   container_name = "nginx"
   container_port = 9000
 
-
-  // See https://github.com/wellcomecollection/platform-infrastructure/tree/master/containers
-  ecr_repo = "760097843905.dkr.ecr.eu-west-1.amazonaws.com/uk.ac.wellcome/nginx_apigw"
-  image    = "${local.ecr_repo}:${var.container_tag}"
+  image = "${var.container_registry}/${var.container_name}:${var.container_tag}"
 }