+++ modules/secret_references/variables.tf
@@ -1,0 +1,4 @@
+variable "secrets" {
+  type    = map(string)
+  default = {}
+}